requirements.txt
discord.py==1.3.4
psycopg2==2.8.5